chore: Suppress compiler warnings and reduce verbose output
- Add compiler flags to suppress common warnings (-Wno-unused-parameter,
  -Wno-sign-compare, -Wno-deprecated-copy)
- Fix sign comparison warnings in SineController and SquareWave
- Remove verbose MaterialManager startup logging

+++ contrib/controller/SquareWave.cpp
@@ -149,7 +149,7 @@ void SquareWave::init()
     }
   }
 
-  for (int i = 0; i < motors.size(); i++)
+  for (size_t i = 0; i < motors.size(); i++)
   {
     _values[i] = _min[i];
   }
